Implements search term highlighting and excerpts
If the content contains the term adds an extract, as if the term does not appear in the title it can be confusing as to why it is shown in the results, especially as WordPress does not return search results using relavency scores by default. Also adds highlighting to make this clearer.